+++ apps/web/src/app/(dashboard)/(management)/@sidebar/locatie/[location]/_components/feedback.tsx
@@ -41,7 +41,7 @@ import {
 import { usePathname, useSearchParams } from "next/navigation";
 import { useFormStatus } from "react-dom";
 import { z } from "zod";
-import { productFeedbackAction } from "~/app/(dashboard)/_actions/feedback";
+import { productFeedbackAction } from "~/actions/send-feedback-action";
 import Spinner from "~/app/_components/spinner";
 
 const feedbackLabels = {
@@ -179,16 +179,16 @@ function Feedback() {
         <DialogTitle>Neem contact op</DialogTitle>
 
         <TabGroup className="mt-4">
-          <TabList className="bg-slate-100 border border-slate-300 rounded-lg w-full grid grid-cols-3 p-[2px]">
-            <Tab className="p-2 data-selected:bg-white flex items-center text-slate-900 gap-x-2 data-selected:opacity-100 opacity-50 justify-center rounded-md text-xs data-selected:shadow-sm">
+          <TabList className="grid grid-cols-3 bg-slate-100 p-[2px] border border-slate-300 rounded-lg w-full">
+            <Tab className="flex justify-center items-center gap-x-2 data-selected:bg-white opacity-50 data-selected:opacity-100 data-selected:shadow-sm p-2 rounded-md text-slate-900 text-xs">
               <BugAntIcon className="size-4 text-slate-400" />
               Bug
             </Tab>
-            <Tab className="p-2 data-selected:bg-white flex items-center text-slate-900 gap-x-2 data-selected:opacity-100 opacity-50 justify-center rounded-md text-xs data-selected:shadow-sm">
+            <Tab className="flex justify-center items-center gap-x-2 data-selected:bg-white opacity-50 data-selected:opacity-100 data-selected:shadow-sm p-2 rounded-md text-slate-900 text-xs">
               <ChatBubbleOvalLeftIconSm className="size-4 text-slate-400" />
               Feedback
             </Tab>
-            <Tab className="p-2 data-selected:bg-white flex items-center text-slate-900 gap-x-2 data-selected:opacity-100 opacity-50 justify-center rounded-md text-xs data-selected:shadow-sm">
+            <Tab className="flex justify-center items-center gap-x-2 data-selected:bg-white opacity-50 data-selected:opacity-100 data-selected:shadow-sm p-2 rounded-md text-slate-900 text-xs">
               <QuestionMarkCircleIcon className="size-4 text-slate-400" />
               Vraag
             </Tab>
